backend/requirements.txt
fastapi[standard]
alembic
pydantic_settings
PyJWT
python-jose[cryptography]
passlib[bcrypt]
requests
redis[hiredis]
chess